refactor: move value objects, use cases and application service location
