+++ restaurant-service/src/main/kotlin/com/munchies/restaurant/domain/valueobject/restaurant/Phone.kt
@@ -1,4 +1,4 @@
-package com.munchies.restaurant.domain.valueobject
+package com.munchies.restaurant.domain.valueobject.restaurant
 
 @JvmInline
 value class Phone private constructor(val value: String) {
