+++ restaurant-service/src/main/kotlin/com/munchies/restaurant/domain/valueobject/restaurant/RestaurantName.kt
@@ -1,4 +1,4 @@
-package com.munchies.restaurant.domain.valueobject
+package com.munchies.restaurant.domain.valueobject.restaurant
 
 @JvmInline
 value class RestaurantName private constructor(val value: String) {
